backend/requirement.txt
asgiref==3.8.1
channels==4.1.0
Django==5.0.4
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
Markdown==3.6
pillow==10.3.0
PyJWT==2.8.0
sqlparse==0.4.4
